Correction in calculating potential density
This is an example of how to calculate potential density from the profiler data. I have implemented it just for Year 1 of the southern ocean lower profiler data, but this will be relevant for all profiler data (so also for SO upper data and for Irminger data) in all years.