Updated SO upper profiler script
The main fix here accounts for the NaN values in the year 1 pressure data. This will also require pulling a new change to the profileIndex function, which was pushed to the OOI_Irminger repository.

Also fixed the problems with reading in the optode_temperature previously fixed in the SO lower profiler script, and temporarily removed the section for calculating potential density.